refactor: insertMarkdown 함수를 useMarkdownEditor 커스텀 훅으로 분리 (#96)
