+++ src/components/markdown/MarkdownEditor.tsx
@@ -1,33 +1,15 @@
-import { useRef, useState } from 'react'
+import { useState } from 'react'
 import {
   EditorHeader,
   EditorTextarea,
   MarkdownExample,
   Preview,
 } from '@/components/markdown'
+import { useMarkdownEditor } from '@/hooks'
 
 export function MarkdownEditor() {
+  const { value, setValue, textareaRef, insertMarkdown } = useMarkdownEditor()
   const [mode, setMode] = useState<'write' | 'preview'>('write')
-  const [value, setValue] = useState('')
-  const textareaRef = useRef<HTMLTextAreaElement | null>(null)
-
-  function insertMarkdown(before: string, after: string = '') {
-    const textarea = textareaRef.current
-    if (!textarea) return
-
-    const start = textarea.selectionStart
-    const end = textarea.selectionEnd
-
-    const selected = value.slice(start, end)
-    const replaced = before + (selected || '') + after
-
-    setValue(value.slice(0, start) + replaced + value.slice(end))
-
-    requestAnimationFrame(() => {
-      textarea.focus()
-      textarea.selectionStart = textarea.selectionEnd = start + replaced.length
-    })
-  }
 
   return (
     <div className="border-custom-gray-200 rounded-lg border">
